Made the live gtk reload more robust

+++ wpgtk/data/reload.py
@@ -2,6 +2,9 @@ import shutil
 import subprocess
 import os
 import tempfile
+import logging
+from .util import setup_log
+setup_log()
 
 from pywal import reload
 
@@ -37,32 +40,57 @@ def openbox():
 
 def gtk3():
     if settings.getboolean("gtk", True):
+        gsettings_theme = None
+        if shutil.which("gsettings") and subprocess.call(["gsettings", "get", "org.gnome.desktop.interface", "gtk-theme"],stdout=subprocess.DEVNULL) == 0:
+            gsettings_theme = subprocess.Popen(["gsettings", "get", "org.gnome.desktop.interface", "gtk-theme"],stdout=subprocess.PIPE).communicate()[0].decode().strip("' \n")
 
-        if shutil.which("xsettingsd"):
-            fd, path = tempfile.mkstemp()
-            try:
-                with os.fdopen(fd, 'w+') as tmp:
-                    tmp.write('Net/ThemeName "FlatColor"\n')
-                    tmp.close()
-                    subprocess.call(
-                        ["timeout", "0.2s", "xsettingsd", "-c", path],
-                        stdout=subprocess.DEVNULL,
-                        stderr=subprocess.DEVNULL
-                    )
-            finally:
-                os.remove(path)
+        refresh_gsettings = "gsettings set org.gnome.desktop.interface gtk-theme '' && sleep 0.1 && gsettings set org.gnome.desktop.interface gtk-theme '%s'" % gsettings_theme
 
-        elif shutil.which("gsettings"):
+        # if gnome-settings-daemon is running, no need to use xsettingsd
+        if subprocess.call(["pgrep", "gsd-xsettings"],stdout=subprocess.DEVNULL) == 0 and gsettings_theme:
+            subprocess.Popen(refresh_gsettings, shell=True)
+            logging.info("Reloaded %s theme via gnome-settings-daemon" % gsettings_theme)
+
+        # no settings daemon is running. So GTK is getting theme info from gtkrc file
+        # So using xsettingd to set the same theme (parsing it from gtkrc)
+        elif shutil.which("xsettingsd") and os.path.isfile(os.path.join(os.environ.get('XDG_CONFIG_HOME'), 'gtk-3.0', 'settings.ini')):
+            import configparser
+            gtkrc = configparser.ConfigParser()
+            gtkrc.read(os.path.join(os.environ.get('XDG_CONFIG_HOME'), 'gtk-3.0', 'settings.ini'))
+            if "Settings" in gtkrc and "gtk-theme-name" in gtkrc["Settings"]:
+                theme_name = gtkrc["Settings"]["gtk-theme-name"]
+                fd, path = tempfile.mkstemp()
+                try:
+                    with os.fdopen(fd, 'w+') as tmp:
+                        tmp.write('Net/ThemeName "'+theme_name+'"\n')
+                        tmp.close()
+                        subprocess.Popen(
+                            ["timeout", "0.2s", "xsettingsd", "-c", path],
+                            stdout=subprocess.DEVNULL,
+                            stderr=subprocess.DEVNULL,
+                        ).communicate() # Don't know why I need to communicate. But without calling this, theme dont update
+                    logging.info("Reloaded theme %s from gtk-3.0/settings.ini using xsettingsd" % theme_name)
+                finally:
+                    os.remove(path)
+        
+        # The system has no known settings daemon installed, but dconf gtk-theme exists, just refreshing its theme
+        # Because user might be using unknown settings daemon
+        elif gsettings_theme:
+            subprocess.Popen(refresh_gsettings, shell=True)
+            logging.warning("No settings daemon found, just refreshing %s theme from gsettings" % gsettings_theme)
+
+
+        elif shutil.which("xsettingsd") and gsettings_theme:
             subprocess.Popen([
-                "gsettings", "set",
-                "org.gnome.desktop.interface", "gtk-theme", "''"
+                "gsettings", "reset",
+                "org.gnome.desktop.interface", "gtk-theme"
             ])
             subprocess.Popen([
                 "gsettings", "set",
                 "org.gnome.desktop.interface", "gtk-theme",
                 "'FlatColor'"
             ])
-
+            logging.warning("No gtk theme is set. So falling back to 'FlatColor' Theme")
 
 def all():
     """Calls all possible reload methods at once."""
